доработал ввод инспекций

+++ Plan/train/views.py
@@ -1,25 +1,40 @@
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import get_object_or_404, redirect, render
 
-from .forms import CasesForm, NewTrainForm, NewMaiForm
-from .models import Cases, Train, DoneMaiDate
-from .utils import page_control
+from .forms import CasesForm, NewTrainForm, NewMaiForm, NewMaiFormFromList
+from .models import Cases, Train, DoneMaiDate, Maintenance
+from .utils import page_control, result_mai_list
 
 PAGE_LIST = 30
 
 
 @login_required
 def train_list(request):
-    cases = Train.objects.all()
-    page_obj = page_control(request, cases, PAGE_LIST)
-    context = {
-        'page_obj': page_obj,
-    }
+    """Вывод списка поездов."""
+
+    trains = Train.objects.all()
+    page_obj = page_control(request, trains, PAGE_LIST)
+    context = {'page_obj': page_obj, }
     return render(request, 'trains/train_list.html', context)
 
 
 @login_required
+def train_create(request):
+    """Создание поезда."""
+
+    form = NewTrainForm(request.POST or None, files=request.FILES or None)
+    if not form.is_valid():
+        return render(request, 'trains/train_create.html', {'form': form})
+    train = form.save(commit=False)
+    train.author = request.user
+    train.save()
+    return redirect('train:train_list')
+
+
+@login_required
 def cases_list(request, train_id):
+    """Вывод списка замечаний на поезде."""
+
     train = get_object_or_404(Train, id=train_id)
     cases = Cases.objects.filter(train=train)
     page_obj = page_control(request, cases, PAGE_LIST)
@@ -32,7 +47,9 @@ def cases_list(request, train_id):
 
 @login_required
 def case_detail(request, case_id):
-    case = Cases.objects.select_related().get(id=case_id)
+    """Редактирование замечания."""
+
+    case = get_object_or_404(Cases, id=case_id)
     form = CasesForm(
         request.POST or None,
         files=request.FILES or None,
@@ -45,13 +62,14 @@ def case_detail(request, case_id):
             instance=case)
 
     context = {'form': form,
-               'case': case,
-               'is_edit': True, }
+               'case': case, }
     return render(request, 'trains/case_detail.html', context)
 
 
 @login_required
 def case_create(request, train_id):
+    """Создание замечания."""
+
     form = CasesForm(request.POST or None, files=request.FILES or None)
     train = get_object_or_404(Train, id=train_id)
     context = {'form': form,
@@ -66,44 +84,50 @@ def case_create(request, train_id):
 
 
 @login_required
-def train_create(request):
-    form = NewTrainForm(request.POST or None, files=request.FILES or None)
-    context = {'form': form}
-    if not form.is_valid():
-        return render(request, 'trains/train_create.html', context)
-    train = form.save(commit=False)
-    train.author = request.user
-    train.save()
-    return redirect('train:train_list')
-
-
-@login_required
 def case_delete(request, case_id, train_id):
+    """Удаление замечания."""
+
     case = Cases.objects.filter(id=case_id)
-    if case.exists():
-        case.delete()
+    case.delete() if case.exists() else None
     return redirect('train:cases_list', train_id)
 
 
 @login_required
 def mai_list(request, train_id):
+    """Вывод списка проведенных и ближайших инспекций."""
+
     train = get_object_or_404(Train, id=train_id)
-    mai = DoneMaiDate.objects.filter(train=train)
-    page_obj = page_control(request, mai, PAGE_LIST)
+    donemai = DoneMaiDate.objects.filter(train=train)
+    main = Maintenance.objects.filter(order=True)
+    result = result_mai_list(main, donemai)
     context = {
-        'page_obj': page_obj,
+        'result': result,
         'train': train,
     }
     return render(request, 'trains/mai_list.html', context)
 
 
 @login_required
+def mai_delete(request, mai_id):
+    """Удаление проведенной инспекции."""
+
+    mai = DoneMaiDate.objects.filter(id=mai_id)
+    if mai.exists():
+        train = mai[0].train
+        mai.delete()
+    return redirect('train:mai_list', train.id)
+
+
+@login_required
 def mai_create(request, train_id):
-    form = NewMaiForm(request.POST or None, files=request.FILES or None)
+    """Создание проведенной инспекции (произвольной)."""
+
+    form = NewMaiForm(
+        request.POST or None,
+        files=request.FILES or None, )
     train = get_object_or_404(Train, id=train_id)
     context = {'form': form,
-               'train': train,
-               }
+               'train': train, }
     if not form.is_valid():
         return render(request, 'trains/mai_create.html', context)
     mai = form.save(commit=False)
@@ -111,3 +135,53 @@ def mai_create(request, train_id):
     mai.train = train
     mai.save()
     return redirect('train:mai_list', train_id)
+
+
+@login_required
+def mai_create_from_list(request, train_id, mai_id):
+    """Создание проведенной инспекции из списка инспекций."""
+
+    mai_type = get_object_or_404(Maintenance, id=mai_id)
+    form = NewMaiFormFromList(
+
+        request.POST or None,
+        files=request.FILES or None,
+    )
+    train = get_object_or_404(Train, id=train_id)
+    context = {'form': form,
+               'train': train,
+               'create': True,
+               'mai_type': mai_type, }
+    if not form.is_valid():
+        return render(request, 'trains/mai_create.html', context)
+    mai_done = form.save(commit=False)
+    mai_done.author = request.user
+    mai_done.train = train
+    mai_done.maintenance = mai_type
+    mai_done.save()
+    return redirect('train:mai_list', train_id)
+
+
+@login_required
+def mai_detail(request, mai_id):
+    """Редактирование проведенной инспекции."""
+
+    mai_done = get_object_or_404(DoneMaiDate, id=mai_id)
+    train = get_object_or_404(Train, id=mai_done.train.id)
+    form = NewMaiFormFromList(
+        request.POST or None,
+        files=request.FILES or None,
+        instance=mai_done, )
+    if form.is_valid():
+        mai_done.save()
+        form = NewMaiFormFromList(
+            request.POST or None,
+            files=request.FILES or None,
+            instance=mai_done)
+        return redirect('train:mai_list', train.id)
+    context = {'form': form,
+               'train': train,
+               'is_edit': True,
+               'mai_done': mai_done,
+               'mai_type': mai_done.maintenance, }
+    return render(request, 'trains/mai_create.html', context)
